tabulon_dxf: Correct Arc translation.
Angles in dxf-rs are in degrees, but angles in Kurbo are in radians,
and the start and end angles are used instead of sweep angles.
These are now correctly calculated for Kurbo use.

Correctly translating the angles dramatically improves performance
because Arcs are now 360 times shorter, thus generate fewer segments.

`Arc` is moved to the set of shapes that are converted
to paths before being transformed, because I was getting
incorrect results from `Affine` multiplication with `kurbo::Arc`.