vello_viewer: don't reproject the scene more than once per redraw.
This deferred reprojection state allows the (expensive) reprojection
to be deferred until after the next frame is drawn.
This is particularly helpful for when input devices deliver events
more frequently than we draw frames.
This does not increase latency significantly, but could be further
refined with a frame pacing controller.