tabulon_dxf: Derive MTEXT alignment from attachment point.
MTEXT derives its default inline alignment from the attachment point,
so I've added that.

Previously I was interpreting `reference_rectangle_width` as the
reflow width of the MTEXT. It turns out that there are circumstances
where this is not interpreted that way.

There is a `column_type` field which affects whether the column width
is respected or not.

tabulon_vello now does attachment points based on the explicit width
of the layout (if provided) to match with how it applies alignment.