Use `sync::Arc` to reference `AnyShape`s, and restructure.
Putting shapes in `Arc` can make a lot of functionality much simpler,
and doesn't (currently) cost a measurable amount of performance.

This also switches `vello_viewer` to debug printing the `Entity` that
was picked, rather than the shape that was picked, this is a step
toward surfacing more interesting information through `tabulon_dxf` to
display on screen.