Introduce managed transforms on `GraphicsBag`.
Managed transforms have a four byte `TransformHandle` which is a lot
more space efficient than an `Affine` on each item.

tabulon_dxf: Expose entity handles for lines.

vello_viewer: Debug print the `Entity` that was picked, rather than
the shape that was picked, this is a step toward surfacing more
interesting information to display on screen.